music_player_app
This code is a music player app in Flutter. we can play the music from internet and using internal as well. In this project I just include the internal of the app. First you give the storage permission then this app will find mp3 audio music files in the internal and finally play the music. I hve used GetX state management technique and MVC architecture.